added ability to edit spells
