+++ src/app/edit/[id]/page.tsx
@@ -1,0 +1,3 @@
+export default function EditSpell() {
+  return <div>EditSpell</div>;
+}
